Fixing the memory consumption problem of mapping db update other small bug fixes

+++ structman_source/structman/scripts/updateMappingDB.py
@@ -62,15 +62,15 @@ def check_instance(config, fromScratch = False):
         
     return fresh_instance
 
-def retrieve_raw_data(config, fromScratch = False):
-    mapping_file_path = retrieve_data_from_uniprot(config, 'idmapping', 'idmapping.dat.gz', fromScratch = fromScratch)
+def retrieve_raw_data(config, raw_files_folder_path, fromScratch = False):
+    mapping_file_path = retrieve_data_from_uniprot(config, raw_files_folder_path, 'idmapping', 'idmapping.dat.gz', fromScratch = fromScratch)
     sequence_file_paths = []
     for seq_file_name in ['uniprot_sprot_varsplic.fasta.gz','uniprot_sprot.fasta.gz','uniprot_trembl.fasta.gz']:
-        sequence_file_paths.append(retrieve_data_from_uniprot(config, 'complete', seq_file_name, fromScratch = fromScratch))
+        sequence_file_paths.append(retrieve_data_from_uniprot(config, raw_files_folder_path, 'complete', seq_file_name, fromScratch = fromScratch))
     return mapping_file_path, sequence_file_paths
 
-def retrieve_data_from_uniprot(config, uniprot_sub_folder, uniprot_file_name, fromScratch = False):
-    file_path = '%s/%s' % (config.tmp_folder, uniprot_file_name)
+def retrieve_data_from_uniprot(config, raw_files_folder_path, uniprot_sub_folder, uniprot_file_name, fromScratch = False):
+    file_path = '%s/%s' % (raw_files_folder_path, uniprot_file_name)
     if os.path.isfile(file_path) and fromScratch:
         os.remove(file_path)
     if not os.path.isfile(file_path):
@@ -86,16 +86,27 @@ def put_seqs_to_database(seq_map, config):
 
     database.update(config, 'UNIPROT', ['Uniprot_Ac','Sequence'], values, mapping_db = True)
 
-def main(config, fromScratch = False):
+def main(config, fromScratch = False, update_mapping_db_keep_raw_files = False):
     #Step 1: Check if mapping SQL DB instance is there, create if not. Recreate for fromScratch mode
     fresh_instance = check_instance(config, fromScratch = fromScratch)
 
     print('Mapping DB instance checked, fresh_instance:', fresh_instance)
 
-    #Step 2: Check for raw files and download if necessary
-    mapping_file_path, seq_file_paths = retrieve_raw_data(config, fromScratch = fromScratch)
+    if update_mapping_db_keep_raw_files:
+        if config.container_version:
+            raw_files_folder_path = '/structman/resources/'
+        else:
+            raw_files_folder_path = config.base_path
+    else:
+        raw_files_folder_path = config.tmp_folder
 
-    print('\nDownloading all raw data files done.\n')
+    #Step 2: Check for raw files and download if necessary
+    mapping_file_path, seq_file_paths = retrieve_raw_data(config, raw_files_folder_path, fromScratch = fromScratch)
+
+    if update_mapping_db_keep_raw_files:
+        print(f'\nDownloading all raw data files done. The files are stored in: {raw_files_folder_path}\n')
+    else:
+        print(f'\nDownloading all raw data files done. The files are temporarily stored in: {raw_files_folder_path}\n')
 
     #Step 3: Update the database
 
@@ -103,7 +114,7 @@ def main(config, fromScratch = False):
     ac_ref_values = []
     ac_ref_nt_values = []
 
-    max_values_at_a_time = 1000000 * config.gigs_of_ram
+    max_values_at_a_time = int(1000000 * config.gigs_of_ram)
 
     with gzip.open(mapping_file_path, 'rb') as f:
         for line in f:
@@ -152,7 +163,7 @@ def main(config, fromScratch = False):
         database.update(config, 'UNIPROT', ['Uniprot_Ac', 'RefSeq_NT'], ac_ref_nt_values, mapping_db = True)
     print('\nDatabase update of RefSeq NTs done.\n')
 
-    max_seqs_at_a_time = 100000 * config.gigs_of_ram
+    max_seqs_at_a_time = int(70000 * config.gigs_of_ram)
 
     for seq_file in seq_file_paths:
         with gzip.open(seq_file, 'rb') as f:
@@ -174,12 +185,13 @@ def main(config, fromScratch = False):
                     seq_map[u_ac] += (line)
             if len(seq_map) > 0:
                 put_seqs_to_database(seq_map, config)
-            
+                seq_map = {}
             print('\nDatabase update of sequences done.\n')
 
-    #Step 4: Remove the raw files
-    os.remove(mapping_file_path)
-    for seq_file in seq_file_paths:
-        os.remove(seq_file)
+    if not update_mapping_db_keep_raw_files:
+        #Step 4: Remove the raw files
+        os.remove(mapping_file_path)
+        for seq_file in seq_file_paths:
+            os.remove(seq_file)
 
-    print('\nRemoving raw data files done.\n')
+        print('\nRemoving raw data files done.\n')
